Move core functionality into separate crate
