+++ Cargo.toml
@@ -1,0 +1,3 @@
+[workspace]
+members = ["dqmj1_rom_util", "src-tauri"]
+resolver = "2"
