+++ dqmj1_rom_util/src/skill_tbl/na.rs
