+++ dqmj1_rom_util/test/data/unnecessary_newlines.dqmj1_asm
